+++ src/pages/mobile/NotificationPage/NotificationPage.style.ts
